+++ golox/interpreter/builtins.go
@@ -1,6 +1,8 @@
 package interpreter
 
 import (
+	"strconv"
+	"strings"
 	"time"
 )
 
@@ -11,7 +13,18 @@ var builtIns = map[string]loxValue{
 	"type": newBuiltInLoxFunction("type", []string{"value"}, func(args []loxValue) loxValue {
 		return loxString(args[0].Type())
 	}),
+	"parseNumber": newBuiltInLoxFunction("parseNumber", []string{"str"}, func(args []loxValue) loxValue {
+		str, ok := args[0].(loxString)
+		if !ok {
+			return newErrorMsgf("expected parseNumber argument to be a %m, got %m", loxTypeString, args[0].Type())
+		}
+		f, err := strconv.ParseFloat(strings.TrimSpace(str.String()), 64)
+		if err != nil {
+			return newErrorMsgf("%q could not be parsed as a %m", str, loxTypeNumber)
+		}
+		return loxNumber(f)
+	}),
 	"error": newBuiltInLoxFunction("error", []string{"msg"}, func(args []loxValue) loxValue {
-		return errorMsg(args[0].String())
+		return newErrorMsg(args[0].String())
 	}),
 }